chore: Moved wiki installation out of wiki container start (#44)
The docker compose entrypoint has a conditional section to perform the
wikimedia installation from CLI and it can get stuck in a loop
preventing the container from ever starting.

Following what Mediawiki does, moved the wiki installation script out of
the entrypoint.

The order now is (from a clean slate):
1. docker compose up -d
   * The db and the wiki web service should both be running
2. docker build --no-cache
   * ensures all the wiki extensions are fetched
   * the installwiki.sh will be copied to the container
3. docker exec <wiki-container-name> /bin/bash /installwiki.sh
   * This will perform some mediawiki CLI installation
   * Replace the LocalSettings.php generated from the CLI run with the
     repo's copy that contains the configured extensions and skins
   * Execute the php maintenance/run update
4. Check the wiki http://localhost:8080/index.php/Special:Version